Add files via upload
This is what I've worked on with some of zuckermans code and just following along in the book. I will probably get it into fully working with scoreboard tonight and then we can start tweeking things to make it more our own project if ya'll are cool with that. Not super familiar with github so I apologize if this creates duplicates. We may have to work on version control.